requirements.txt
fastapi~=0.115.6
uvicorn[standard]~=0.33.0
jinja2
SQLAlchemy~=2.0.36
alembic~=1.13.3
python-dotenv~=1.0.1
redis~=5.2.1
pydantic~=2.7.3
passlib~=1.7.4